+++ AnnoMapEditor/UI/Controls/IslandProperties.xaml
@@ -4,6 +4,7 @@
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
              xmlns:local="clr-namespace:AnnoMapEditor.UI.Controls"
+             xmlns:converters="clr-namespace:AnnoMapEditor.UI.Converters"
              xmlns:maptemplates="clr-namespace:AnnoMapEditor.MapTemplates"
              d:DataContext="{d:DesignInstance Type=maptemplates:Island}"
              mc:Ignorable="d" 
@@ -13,7 +14,7 @@
             <ResourceDictionary.MergedDictionaries>
                 <ResourceDictionary Source="../Styles.xaml" />
                 <ResourceDictionary>
-                    <local:ObjectToVisibility x:Key="CollapsedOnNull" />
+                    <converters:ObjectToVisibility x:Key="CollapsedOnNull" />
                 </ResourceDictionary>
             </ResourceDictionary.MergedDictionaries>
         </ResourceDictionary>
