+++ AnnoMapEditor/UI/Converters/BoolToVisibility.cs
@@ -1,0 +1,34 @@
+﻿using System;
+using System.Globalization;
+using System.Windows;
+using System.Windows.Data;
+
+namespace AnnoMapEditor.UI.Converters
+{
+    [ValueConversion(typeof(bool), typeof(Visibility))]
+    public class BoolToVisibility : IValueConverter
+    {
+        public Visibility OnTrue { get; set; }
+        public Visibility OnFalse { get; set; }
+
+
+        public BoolToVisibility()
+        {
+            OnTrue = Visibility.Visible;
+            OnFalse = Visibility.Collapsed;
+        }
+
+
+        public object? Convert(object value, Type targetType, object parameter, CultureInfo culture)
+        {
+            if (value is not bool boolean)
+                return null;
+            return boolean ? OnTrue : OnFalse;
+        }
+
+        public object? ConvertBack(object value, Type targetType, object parameter, CultureInfo culture)
+        {
+            throw new NotImplementedException();
+        }
+    }
+}
