+++ AnnoMapEditor/UI/Converters/NegateDoubleConverter.cs
@@ -1,9 +1,7 @@
 ﻿using System;
-using System.Collections.Generic;
+
 using System.Globalization;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+
 using System.Windows.Data;
 
 namespace AnnoMapEditor.UI.Converters
